moving api request to its own file in services
